Move RAC collections implementation into new package
