+++ packages/@react-aria/collections/README.md
@@ -1,0 +1,3 @@
+# @react-aria/collections
+
+This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
