+++ packages/@react-aria/collections/src/BaseCollection.ts
@@ -1,0 +1,212 @@
+/*
+ * Copyright 2024 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {Collection as ICollection, Node} from 'react-stately';
+import {Key} from '@react-types/shared';
+import {ReactElement, ReactNode} from 'react';
+
+export type Mutable<T> = {
+  -readonly[P in keyof T]: T[P]
+}
+
+/** An immutable object representing a Node in a Collection. */
+export class NodeValue<T> implements Node<T> {
+  readonly type: string;
+  readonly key: Key;
+  readonly value: T | null = null;
+  readonly level: number = 0;
+  readonly hasChildNodes: boolean = false;
+  readonly rendered: ReactNode = null;
+  readonly textValue: string = '';
+  readonly 'aria-label'?: string = undefined;
+  readonly index: number = 0;
+  readonly parentKey: Key | null = null;
+  readonly prevKey: Key | null = null;
+  readonly nextKey: Key | null = null;
+  readonly firstChildKey: Key | null = null;
+  readonly lastChildKey: Key | null = null;
+  readonly props: any = {};
+  readonly render?: (node: Node<any>) => ReactElement;
+
+  constructor(type: string, key: Key) {
+    this.type = type;
+    this.key = key;
+  }
+
+  get childNodes(): Iterable<Node<T>> {
+    throw new Error('childNodes is not supported');
+  }
+
+  clone(): NodeValue<T> {
+    let node: Mutable<NodeValue<T>> = new NodeValue(this.type, this.key);
+    node.value = this.value;
+    node.level = this.level;
+    node.hasChildNodes = this.hasChildNodes;
+    node.rendered = this.rendered;
+    node.textValue = this.textValue;
+    node['aria-label'] = this['aria-label'];
+    node.index = this.index;
+    node.parentKey = this.parentKey;
+    node.prevKey = this.prevKey;
+    node.nextKey = this.nextKey;
+    node.firstChildKey = this.firstChildKey;
+    node.lastChildKey = this.lastChildKey;
+    node.props = this.props;
+    node.render = this.render;
+    return node;
+  }
+}
+
+/**
+ * An immutable Collection implementation. Updates are only allowed
+ * when it is not marked as frozen. This can be subclassed to implement
+ * custom collection behaviors.
+ */
+export class BaseCollection<T> implements ICollection<Node<T>> {
+  private keyMap: Map<Key, NodeValue<T>> = new Map();
+  private firstKey: Key | null = null;
+  private lastKey: Key | null = null;
+  private frozen = false;
+
+  get size() {
+    return this.keyMap.size;
+  }
+
+  getKeys() {
+    return this.keyMap.keys();
+  }
+
+  *[Symbol.iterator]() {
+    let node: Node<T> | undefined = this.firstKey != null ? this.keyMap.get(this.firstKey) : undefined;
+    while (node) {
+      yield node;
+      node = node.nextKey != null ? this.keyMap.get(node.nextKey) : undefined;
+    }
+  }
+
+  getChildren(key: Key): Iterable<Node<T>> {
+    let keyMap = this.keyMap;
+    return {
+      *[Symbol.iterator]() {
+        let parent = keyMap.get(key);
+        let node = parent?.firstChildKey != null ? keyMap.get(parent.firstChildKey) : null;
+        while (node) {
+          yield node as Node<T>;
+          node = node.nextKey != null ? keyMap.get(node.nextKey) : undefined;
+        }
+      }
+    };
+  }
+
+  getKeyBefore(key: Key) {
+    let node = this.keyMap.get(key);
+    if (!node) {
+      return null;
+    }
+
+    if (node.prevKey != null) {
+      node = this.keyMap.get(node.prevKey);
+
+      while (node && node.type !== 'item' && node.lastChildKey != null) {
+        node = this.keyMap.get(node.lastChildKey);
+      }
+
+      return node?.key ?? null;
+    }
+
+    return node.parentKey;
+  }
+
+  getKeyAfter(key: Key) {
+    let node = this.keyMap.get(key);
+    if (!node) {
+      return null;
+    }
+
+    if (node.type !== 'item' && node.firstChildKey != null) {
+      return node.firstChildKey;
+    }
+
+    while (node) {
+      if (node.nextKey != null) {
+        return node.nextKey;
+      }
+
+      if (node.parentKey != null) {
+        node = this.keyMap.get(node.parentKey);
+      } else {
+        return null;
+      }
+    }
+
+    return null;
+  }
+
+  getFirstKey() {
+    return this.firstKey;
+  }
+
+  getLastKey() {
+    let node = this.lastKey != null ? this.keyMap.get(this.lastKey) : null;
+    while (node?.lastChildKey != null) {
+      node = this.keyMap.get(node.lastChildKey);
+    }
+
+    return node?.key ?? null;
+  }
+
+  getItem(key: Key): Node<T> | null {
+    return this.keyMap.get(key) ?? null;
+  }
+
+  at(): Node<T> {
+    throw new Error('Not implemented');
+  }
+
+  clone(): this {
+    // We need to clone using this.constructor so that subclasses have the right prototype.
+    // TypeScript isn't happy about this yet.
+    // https://github.com/microsoft/TypeScript/issues/3841
+    let Constructor: any = this.constructor;
+    let collection: this = new Constructor();
+    collection.keyMap = new Map(this.keyMap);
+    collection.firstKey = this.firstKey;
+    collection.lastKey = this.lastKey;
+    return collection;
+  }
+
+  addNode(node: NodeValue<T>) {
+    if (this.frozen) {
+      throw new Error('Cannot add a node to a frozen collection');
+    }
+
+    this.keyMap.set(node.key, node);
+  }
+
+  removeNode(key: Key) {
+    if (this.frozen) {
+      throw new Error('Cannot remove a node to a frozen collection');
+    }
+
+    this.keyMap.delete(key);
+  }
+
+  commit(firstKey: Key | null, lastKey: Key | null, isSSR = false) {
+    if (this.frozen) {
+      throw new Error('Cannot commit a frozen collection');
+    }
+
+    this.firstKey = firstKey;
+    this.lastKey = lastKey;
+    this.frozen = !isSSR;
+  }
+}
