+++ packages/react-aria-components/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@internationalized/date": "^3.5.4",
     "@internationalized/string": "^3.2.3",
+    "@react-aria/collections": "3.0.0-alpha.1",
     "@react-aria/color": "3.0.0-beta.33",
     "@react-aria/dnd": "^3.6.1",
     "@react-aria/focus": "^3.17.1",
