+++ packages/react-aria-components/src/Table.tsx
@@ -1,8 +1,9 @@
 import {AriaLabelingProps, HoverEvents, Key, LinkDOMProps} from '@react-types/shared';
-import {BaseCollection, Collection, CollectionBuilder, CollectionContext, CollectionProps, CollectionRendererContext, createBranchComponent, createLeafComponent, ItemRenderProps, NodeValue, useCachedChildren, useCollectionChildren} from './Collection';
+import {BaseCollection, Collection, CollectionBuilder, createBranchComponent, createLeafComponent, NodeValue, useCachedChildren} from '@react-aria/collections';
 import {buildHeaderRows, TableColumnResizeState} from '@react-stately/table';
 import {ButtonContext} from './Button';
 import {CheckboxContext} from './RSPContexts';
+import {CollectionProps, CollectionRendererContext, ItemRenderProps, useClearCollectionRenderer} from './Collection';
 import {ColumnSize, ColumnStaticSize, TableCollection as ITableCollection, TableProps as SharedTableProps} from '@react-types/table';
 import {ContextValue, DEFAULT_SLOT, DOMProps, Provider, RenderProps, ScrollableProps, SlotProps, StyleProps, StyleRenderProps, useContextProps, useRenderProps} from './utils';
 import {DisabledBehavior, DraggableCollectionState, DroppableCollectionState, MultipleSelectionState, Node, SelectionBehavior, SelectionMode, SortDirection, TableState, useMultipleSelectionState, useTableColumnResizeState, useTableState} from 'react-stately';
@@ -563,11 +564,11 @@ export const TableHeader =  /*#__PURE__*/ createBranchComponent(
       </THead>
     );
   },
-  props => useCollectionChildren({
-    children: props.children,
-    items: props.columns,
-    dependencies: props.dependencies
-  })
+  props => (
+    <Collection dependencies={props.dependencies} items={props.columns}>
+      {props.children}
+    </Collection>
+  )
 );
 
 function TableHeaderRow({item}: {item: GridNode<any>}) {
@@ -730,7 +731,7 @@ export const Column = /*#__PURE__*/ createLeafComponent('column', (props: Column
       data-allows-sorting={column.props.allowsSorting || undefined}
       data-sort-direction={state.sortDescriptor?.column === column.key ? state.sortDescriptor.direction : undefined}>
       <ColumnResizerContext.Provider value={{column, triggerRef: ref}}>
-        {renderProps.children}
+        {useClearCollectionRenderer(renderProps.children)}
       </ColumnResizerContext.Provider>
     </TH>
   );
@@ -1100,20 +1101,10 @@ export const Row = /*#__PURE__*/ createBranchComponent(
   },
   props => {
     let dependencies = [props.value].concat(props.dependencies);
-    let children = useCollectionChildren({
-      dependencies,
-      children: props.children,
-      items: props.columns,
-      idScope: props.id
-    });
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    let ctx = useMemo(() => ({idScope: props.id, dependencies}), [props.id, ...dependencies]);
-
     return (
-      <CollectionContext.Provider value={ctx}>
-        {children}
-      </CollectionContext.Provider>
+      <Collection dependencies={dependencies} items={props.columns} idScope={props.id}>
+        {props.children}
+      </Collection>
     );
   }
 );
@@ -1190,7 +1181,7 @@ export const Cell = /*#__PURE__*/ createLeafComponent('cell', (props: CellProps,
       data-focused={isFocused || undefined}
       data-focus-visible={isFocusVisible || undefined}
       data-pressed={isPressed || undefined}>
-      {renderProps.children}
+      {useClearCollectionRenderer(renderProps.children)}
     </TD>
   );
 });
